Preparation for x64 on Windows.

+++ gc_menu.c
@@ -631,9 +631,9 @@ void PMenu_Update(edict_t *ent)
 		}
 		sprintf(string + strlen(string), "yv %d ", 32 + i * 8);
 		if (p->align == PMENU_ALIGN_CENTER)
-			x = 196/2 - strlen(line)*4 + 64;
+			x = 196/2 - (int)strlen(line)*4 + 64;
 		else if (p->align == PMENU_ALIGN_RIGHT)
-			x = 64 + (196 - strlen(line)*8);
+			x = 64 + (196 - (int)strlen(line)*8);
 		else
 			x = 64;
